MemoryRAG/requirements.txt
# OpenAI ja Azure
openai>=1.0.0
azure-ai-ml>=1.0.0
azure-identity>=1.0.0

# Muut riippuvuudet
python-dotenv>=0.19.0
aiohttp>=3.8.0
pandas>=2.0.0
pypdf>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.9.3
requests>=2.25.1

# Testaus
pytest>=7.0.0
pytest-asyncio>=0.21.0  # Async testejä varten